backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.0
python-multipart==0.0.12
jinja2==3.1.4
openpyxl==3.1.5
boto3==1.35.0
azure-identity==1.19.0
azure-mgmt-resource==23.2.0
azure-mgmt-authorization==4.0.0
azure-mgmt-monitor==6.0.0
azure-mgmt-network==27.0.0
azure-mgmt-compute==33.0.0
azure-mgmt-storage==21.2.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-sql==3.0.0
azure-mgmt-security==7.0.0
azure-mgmt-web==7.3.0
azure-mgmt-policyinsights==1.0.0
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-advisor==9.0.0
azure-mgmt-resourcegraph==8.0.0
azure-mgmt-automation==1.0.0
azure-mgmt-loganalytics==13.1.1
azure-mgmt-securityinsight==1.0.0
requests>=2.31.0
google-cloud-asset==3.27.0
google-cloud-iam==2.16.0
google-cloud-logging==3.11.0
google-cloud-resource-manager==1.13.0
google-cloud-kms==3.1.0
google-cloud-compute==1.20.0
google-cloud-org-policy>=1.16.0
google-cloud-storage>=2.14.0
google-cloud-securitycenter>=1.30.0
google-cloud-monitoring>=2.19.0
google-cloud-container>=2.40.0
google-cloud-bigquery>=3.17.0
google-cloud-os-config>=1.16.0
google-cloud-binary-authorization>=1.9.0
google-cloud-recommender>=2.15.0
google-cloud-dns>=0.35.0
google-auth>=2.28.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
authlib==1.3.2
httpx>=0.27.0
itsdangerous==2.2.0
